use HashMap instead of ConcurrentHashMap
