+++ core/src/main/java/com/distkv/dst/core/operatorImpl/DstDictImpl.java
@@ -2,7 +2,7 @@ package com.distkv.dst.core.operatorImpl;
 
 import com.distkv.dst.core.operatorset.DstDict;
 import com.distkv.dst.core.DstMapInterface;
-import com.distkv.dst.core.DstConcurrentHashMapImpl;
+import com.distkv.dst.core.HashMapImpl;
 import java.util.Map;
 
 public class DstDictImpl implements DstDict {
@@ -10,7 +10,7 @@ public class DstDictImpl implements DstDict {
   private DstMapInterface<String, Map<String, String>> dictMap;
 
   public DstDictImpl() {
-    this.dictMap = new DstConcurrentHashMapImpl<>();
+    this.dictMap = new HashMapImpl<>();
   }
 
   @Override
