+++ core/src/main/java/com/distkv/dst/core/operatorImpl/DstListImpl.java
@@ -3,7 +3,7 @@ package com.distkv.dst.core.operatorImpl;
 import com.distkv.dst.common.exception.DstListIndexOutOfBoundsException;
 import com.distkv.dst.core.operatorset.DstList;
 import com.distkv.dst.core.DstMapInterface;
-import com.distkv.dst.core.DstConcurrentHashMapImpl;
+import com.distkv.dst.core.HashMapImpl;
 import com.distkv.dst.common.exception.KeyNotFoundException;
 import com.distkv.dst.common.utils.Status;
 import java.util.ArrayList;
@@ -15,7 +15,7 @@ public class DstListImpl implements DstList {
   private DstMapInterface<String, ArrayList<String>> listMap;
 
   public DstListImpl() {
-    this.listMap = new DstConcurrentHashMapImpl<>();
+    this.listMap = new HashMapImpl<>();
   }
 
   @Override
